requirements.txt
flask
openai
faiss-cpu
numpy
line-bot-sdk
requests
